Move frameReader to its own file
