+++ SnackBar/SnackBar/Helpers/View+Extension.swift
@@ -24,28 +24,3 @@ public extension View {
     }
 }
 
-// MARK: - Get frame size
-extension View {
-    func readFrame(in coordinateSpace: CoordinateSpace = .global, for reader: Binding<CGRect>) -> some View {
-        background(
-            GeometryReader { geometry in
-                Color.clear
-                    .preference(
-                        key: FramePreferenceKey.self,
-                        value: geometry.frame(in: coordinateSpace)
-                    )
-                    .onPreferenceChange(FramePreferenceKey.self, perform: { value in
-                        reader.wrappedValue = value
-                    })
-            }
-        )
-    }
-}
-
-private struct FramePreferenceKey: PreferenceKey {
-    static var defaultValue: CGRect = .zero
-    
-    static func reduce(value: inout CGRect, nextValue: () -> CGRect) {
-        value = nextValue()
-    }
-}
